requirements.txt
# API y HTTP
requests==2.31.0
aiohttp==3.9.1

# Configuración
python-dotenv==1.0.0

# Utilidades
python-dateutil==2.8.2

# Logging y debugging
colorlog==6.7.0

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0